+++ index.template.php
@@ -173,7 +173,7 @@ function template_html_below()
  */
 function template_theme_header()
 {
-	global $scripturl, $context, $settings;
+	global $scripturl, $context;
 
 	echo '
 	<header>
@@ -531,7 +531,7 @@ function template_button_strip($button_strip, $direction = '', $strip_options = 
 
 			$button = '
 				<a class="button normal_button_strip_' . $key . (!empty($value['active']) ? ' active' : '') . (isset($value['class']) ? ' ' . $value['class'] : '') . (!empty($value['sub_buttons']) ? ' buttonlist_sub' : '') . '" ' . (!empty($value['url']) ? 'href="' . $value['url'] . '"' : 'href="javascript:void(0);"') . ' ' . (isset($value['custom']) ? ' ' . $value['custom'] : '') . '>
-					' . themecustoms_icon('fa fa-' . (!empty($value['icon']) ? $value['icon'] : $value['text'])) . '
+					<span class="main_icons ' . (!empty($value['icon']) ? $value['icon'] : $value['text']) . '"></span>
 					<span>' . $txt[$value['text']] . '</span>
 				</a>';