+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### 2.1 - 27 June 2024
+- ![New Feature](https://smftricks.com/assets/changelog/tag--plus.png) Changed the memberlist layout.
+- ![Improvement](https://smftricks.com/assets/changelog/tag--pencil.png) Code improvements for 'Theme Customs'
+---
 #### 2.0 - 09 April 2024
 - ![New Feature](https://smftricks.com/assets/changelog/tag--plus.png) Initial Release for SMF 2.1.x